JwtTokenInterceptor, methods for API calls and Tag-based tasks filtering
JwtTokenInterceptor will automatically add the JWT token as the Authorization header in API HTTP requests for logged in users.
Added all the API method calls in OnlineJudgeApi.
Upon clicking a tag from the navigation drawer, the task list from the home page is filtered now to only contain tasks tagged as the selected tag.
Added a title property in the main view model that reflects the title in the top action bar.